requirements.txt
pandas
requests
dash
dash_bootstrap_components
plotly.express
psycopg2-binary
datetime
matplotlib